+++ packages/fast-tooling-react/app/pages/form-and-navigation.tsx
@@ -152,6 +152,7 @@ class FormAndNavigationTestPage extends React.Component<{}, FormTestPageState> {
                         >
                             {JSON.stringify(this.state.data, null, 2)}
                         </pre>
+                        <pre>{this.state.dataLocation}</pre>
                     </div>
                     <div
                         style={{
@@ -176,6 +177,8 @@ class FormAndNavigationTestPage extends React.Component<{}, FormTestPageState> {
                 schema={this.state.schema}
                 childOptions={this.getChildOptions()}
                 onLocationUpdate={this.handleLocationOnChange}
+                dataLocation={this.state.dataLocation}
+                onChange={this.handleDataOnChange}
             />
         );
     }
@@ -282,6 +285,13 @@ class FormAndNavigationTestPage extends React.Component<{}, FormTestPageState> {
      */
     private handleLocationOnChange = (dataLocation: string): void => {
         this.setState({
+            dataLocation,
+        });
+    };
+
+    private handleDataOnChange = (data: any, dataLocation: string): void => {
+        this.setState({
+            data,
             dataLocation,
         });
     };